Code organization
- collector client code collected together
- interfaces moved to be next to their implementations
