+++ collector_client_thrift.go
